requirements.txt
streamlit
google-cloud-aiplatform>=1.38
google-auth
python-docx
PyPDF2
tiktoken